DATA_GENERATOR/datGen-master/requirements.txt
Flask==2.2.2
flask_restx==1.0.3
flask_sqlalchemy==3.0.2
flask-cors==3.0.10
names==0.3.0
okta_jwt_verifier==0.2.3
pandas==1.5.1
python-dotenv==0.21.0
requests==2.28.1
